cache all tags (not just cuda), move everything to config/platform.go, comments
